Fix simtest build on MacOS X.
On mac stdlib.h includes signal.h which conflicts with the
variable of the same name in this file. Unfortunately that
makes it something of a reserved identifier in portable
code, but the only stdlib.h call this file actually uses
is exit(). Returning from main() is equivalent, so we can
just remove the stdlib.h include instead.